Enforce single Cognitive Twin memory path
